Added payment Logic for community items and fixed a chrome issue.
Added logic to activate item if paid in full and to logic to make
payments after item is set.
Chrome does not update in community tab if item picked from tab not from
dropdown then it fails fixed

+++ src/cs499/object/CommunityItem.java
@@ -29,7 +29,7 @@ public class CommunityItem extends Item{
 	
 	private String activationLimitDate;
 	
-	private long foreignId;
+	private int foreignId;
 
 	public int getPaid() {
 		return paid;
@@ -55,11 +55,11 @@ public class CommunityItem extends Item{
 		this.activationLimitDate = activationLimitDate;
 	}
 
-	public long getForeignId() {
+	public int getForeignId() {
 		return foreignId;
 	}
 
-	public void setForeignId(long foreignId) {
+	public void setForeignId(int foreignId) {
 		this.foreignId = foreignId;
 	}
